Dodanie pierwszych testów dla visitora && małe zmiany w kodzie

+++ src/main/java/Bank/Visitor/Visitable.java
@@ -2,5 +2,6 @@ package Bank.Visitor;
 
 public interface Visitable {
 
-    double accept(Visitor visitor);
+    int accept(Visitor visitor);
+
 }
